Attack on ships changes color of grid tile to red
