+++ src/game.js
@@ -1,5 +1,5 @@
 import player from "./player";
-import { makeBoard, updateBoard } from "./DOM";
+import { makeBoard, updateBoard, attackListener } from "./DOM";
 import { ship } from "./ships";
 
 const populateBoard = (plyr) => {
@@ -13,22 +13,6 @@ const populateBoard = (plyr) => {
   plyr.brd.placeShip(cruiser, "vertical", "D", 2);
   plyr.brd.placeShip(submarine, "vertical", "E", 5);
   plyr.brd.placeShip(destroyer, "horizontal", "H", 1);
-  console.log(plyr.brd.board);
-};
-
-const attackListener = (p1, p2) => {
-  const boardArea = document.querySelector(".board-area");
-  boardArea.addEventListener("click", (e) => {
-    if (p1.getTurn() && e.target.hasAttribute("truedata-coords")) {
-      const coordsToAttack = e.target.getAttribute("truedata-coords");
-      p1.attackBoard(p2, coordsToAttack[0], coordsToAttack[1]);
-      console.log(p2.brd.board);
-    } else if (p2.getTurn() && e.target.hasAttribute("falsedata-coords")) {
-      const coordsToAttack = e.target.getAttribute("falsedata-coords");
-      p2.attackBoard(p1, coordsToAttack[0], coordsToAttack[1]);
-      console.log(p1.brd.board);
-    }
-  });
 };
 
 const runGame = (p1, p2) => {
